[Version] App: 2.1.2 beta 3 Database: V5 [Features] Add enableLosePoints mode && multipleChoicesQuestionsCheckingStrategy in GradingSetting [Fix] DynamicValidator still throws exception when ignoreMissingField = true && the field is missing in some cases Wrong readonly display("Lock" icon) in QuestionEditor.vue User submissions and scores missing in ExamRecordDetailView.vue when the user submits null answer Wrong "change question author" & "change question group author" permission request when update questions UpdateQuestionsAction sending unuseful messages [Improve] Auto logout

+++ src/main/java/indi/etern/checkIn/entities/question/impl/QuestionGroupAnswer.java
@@ -37,8 +37,8 @@ public class QuestionGroupAnswer extends Answer<QuestionGroup, List<SingleQuesti
                     .filter(answer ->
                             answer.check().checkedResultType() == CheckedResultType.CORRECT)
                     .count();
-            double score = correctCount * singleMaxScore;
-            double maxScore = singleMaxScore * maxCount;
+            float score = correctCount * singleMaxScore;
+            float maxScore = singleMaxScore * maxCount;
             CheckedResultType checkedResultType;
             if (maxScore == score) {
                 checkedResultType = CheckedResultType.CORRECT;
